+++ src/extension.ts
@@ -1,9 +1,9 @@
 import * as vscode from "vscode";
 import * as vscodeUtil from "./vscodeUtil";
 import { DecorateManager } from "./decoration";
-import { open } from "fs";
 
 export function activate(context: vscode.ExtensionContext) {
+  const channel = vscode.window.createOutputChannel("DecorateKeyword");
   const dm: DecorateManager = new DecorateManager();
 
   async function readData() {
@@ -39,11 +39,8 @@ export function activate(context: vscode.ExtensionContext) {
 
   async function read() {
     await readData();
-    if (dm.autoDecorate === true) {
-      for (const editor of vscode.window.visibleTextEditors) {
-        dm.undecorate(editor.document);
-        dm.decorate(editor.document);
-      }
+    for (const editor of vscode.window.visibleTextEditors) {
+      dm.keep(editor.document);
     }
   }
 
@@ -51,14 +48,23 @@ export function activate(context: vscode.ExtensionContext) {
     await dm.open();
   }
 
+  async function showLanguageIds() {
+    const languages = await vscode.languages.getLanguages();
+    for (const language of languages) {
+      channel.appendLine(language);
+    }
+    channel.show();
+  }
+
   vscodeUtil.registerCommand(context, "decorate-keyword.read", read);
   vscodeUtil.registerCommand(context, "decorate-keyword.open", open);
   vscodeUtil.registerCommand(context, "decorate-keyword.decorate", decorate);
   vscodeUtil.registerCommand(context, "decorate-keyword.undecorate", undecorate);
   vscodeUtil.registerCommand(context, "decorate-keyword.toggle", toggle);
-  vscodeUtil.registerCommand(context, "decorate-keyword.info", () => {
-    console.log(dm.info());
-  });
+  vscodeUtil.registerCommand(context, "decorate-keyword.showLanguageIds", showLanguageIds);
+  // vscodeUtil.registerCommand(context, "decorate-keyword.info", () => {
+  //   console.log(dm.info());
+  // });
 
   vscode.workspace.onDidChangeConfiguration(
     (event) => {
@@ -93,11 +99,7 @@ export function activate(context: vscode.ExtensionContext) {
         return;
       }
       for (const editor of editors) {
-        const redecorate = dm.redecorate(editor.document);
-        if (redecorate === false && dm.autoDecorate === true) {
-          dm.undecorate(editor.document);
-          dm.decorate(editor.document);
-        }
+        dm.keep(editor.document);
       }
     },
     null,
@@ -106,11 +108,7 @@ export function activate(context: vscode.ExtensionContext) {
 
   vscode.workspace.onDidSaveTextDocument(
     (document) => {
-      const redecorate = dm.redecorate(document);
-      if (redecorate === false && dm.autoDecorate === true) {
-        dm.undecorate(document);
-        dm.decorate(document);
-      }
+      dm.keep(document);
     },
     null,
     context.subscriptions